taking same pattern from prompt and putting it over on images
